// hololive-card/enum.ts
export enum HololiveCardType {
    Oshi = "推しホロメン",
    Holomen = "ホロメン",
    Support = "サポート",
    Cheer = "エール",
    Unknown = "Unknown"
}

// TODO: add all rarities
export enum HololiveCardRarity {
    OSR = "OSR",
    U = "U",
    RR = "RR",
    R = "R",
    C = "C",
    OC = "OC",
    SEC = "SEC",
    OUR = "OUR",
    UR = "UR",
    SY = "SY",
    SR = "SR",
    S = "S",
    P = "P",
    Unknown = "Unknown"
}

export enum HololiveCardAttribute {
    White = "白",
    Red = "赤",
    Blue = "青",
    Green = "緑",
    Yellow = "黄",
    Purple = "紫",
    None = "◇",
    Unknown = "Unknown"
}

export enum HololiveOshiSkillType {
    Oshi = "推しスキル",
    SpOshi = "SP推しスキル",
    Unknown = "Unknown"
}

export enum HololiveHolomenBloomLevel {
    Debut = "Debut",
    First = "1st",
    Second = "2nd",
    Spot = "Spot",
    Unknown = "Unknown"
}

export enum HololiveHolomenPassiveSkillType {
    CollabEffect = "コラボエフェクト",
    Gift = "ギフト",
    BloomEffect = "ブルームエフェクト",
    Unknown = "Unknown"
}

export enum HololiveSupportType {
    Event = "イベント",
    Fan = "ファン",
    Tool = "ツール",
    Mascot = "マスコット",
    Item = "アイテム",
    Staff = "スタッフ",
    Unknown = "Unknown"
}

export enum HololiveMerchandiseType {
    Boosters = "Boosters",
    Decks = "Decks",
    Unknown = "Unknown"
}